chore(plasma-core): move helpers to __helpers
