+++ packages/plasma-core/.gitignore
@@ -8,3 +8,4 @@
 /index.js*
 /index.d.ts*
 /es
+/__helpers
