+++ packages/plasma-core/src/__helpers/IconPlaceholder.tsx
